fix: Update docs to use mdx and link to compound component doc
