+++ modules/_labs/pagination/react/stories/pagination.stories.mdx
@@ -1,0 +1,105 @@
+import {Meta} from '@storybook/addon-docs/blocks';
+
+import {Pagination} from '@workday/canvas-kit-labs-react-pagination';
+import {StepControls} from './examples/StepControls';
+import {CustomRange} from './examples/CustomRange';
+import {JumpControls} from './examples/JumpControls';
+import {GoToForm} from './examples/GoToForm';
+import {ShowAdditionalDetails} from './examples/ShowAdditionalDetails';
+import {HoistedModel} from './examples/HoistedModel';
+import {RTL} from './examples/RTL';
+
+<Meta title="Labs/Pagination/React" component={Pagination} />
+
+# Canvas Kit Labs Pagination
+
+`Pagination` is a compound component for handling navigation between pages in a range.
+
+## Installation
+
+```sh
+yarn add @workday/canvas-kit-labs-react-pagination
+```
+
+## Examples
+
+<br />
+
+### Basic Usage
+
+<br />
+
+#### Step Controls
+
+In this example, the component uses step controls (`Pagination.StepToPreviousButton` and
+`Pagination.StepToNextButton`) that allow you to move to the next page and previous pages.
+
+<ExampleCodeBlock code={StepControls} />
+
+---
+
+#### Jump Controls
+
+This example adds "jump" controls (`Pagination.JumpToFirstButton` and `Pagination.JumpToLastButton`)
+that allow you to skip to the first and last items in the range.
+
+<ExampleCodeBlock code={JumpControls} />
+
+---
+
+#### GoTo Form
+
+This example adds a form (`Pagination.GoToForm`) that allows you to skip to a specific page within
+the range.
+
+<ExampleCodeBlock code={GoToForm} />
+
+---
+
+#### Additional Details
+
+This example adds a visible section (`Pagination.AdditionalDetails`). It is an `aria-live` region
+that announces the current page update to screen readers. Because of that, it's important to
+**always†** include it in your `Pagination` component.
+
+In the case where you would also have multiple `Pagination` components sharing the same state and
+you'd like to keep the `AdditionalDetails` component on multiple, you will need to set
+`shouldAnnounceToScreenReader` to `false` on all but one component to prevent announcement.
+
+> **†** _The only exception to this rule is when you have multiple `Pagination` components that are
+> sharing the same state and rendered on the page. You can then safely remove all but one of the
+> `AdditionalDetails` sections. This will prevent a screenreader from announcing updates multiple
+> times to a user._
+
+<ExampleCodeBlock code={ShowAdditionalDetails} />
+
+---
+
+#### RTL (Right-to-Left)
+
+This example shows how the component supports right-to-left languages.
+
+<ExampleCodeBlock code={RTL} />
+
+---
+
+#### Custom Range
+
+This example uses a custom range that allows you to control the width of the component.
+
+<ExampleCodeBlock code={CustomRange} />
+
+---
+
+### Advanced Usage
+
+<br />
+
+#### Hoisted Model Pattern
+
+If you want the `Pagination` component to handle its state and events internally and hook into page
+change events, the examples above should be sufficient. However, if you need external access to the
+model, you can use the hoisted model pattern. You can create a model outside of the component with
+the `usePaginationModel` hook.
+
+<ExampleCodeBlock code={HoistedModel} />
