--- modules/_labs/pagination/react/stories/stories.tsx
@@ -1,303 +1,0 @@
-/// <reference path="../../../../../typings.d.ts" />
-import * as React from 'react';
-import withReadme from 'storybook-readme/with-readme';
-import {CanvasProvider, ContentDirection} from '@workday/canvas-kit-react-common';
-
-import README from '../README.md';
-
-import {
-  Pagination,
-  getLastPage,
-  getVisibleResultsMax,
-  getVisibleResultsMin,
-  usePaginationModel,
-} from '../lib/Pagination';
-
-export default {
-  title: 'Labs/Pagination/React',
-  decorators: [withReadme(README)],
-};
-
-export const StepControls = () => {
-  const resultCount = 10;
-  const totalCount = 100;
-  const lastPage = getLastPage(resultCount, totalCount);
-
-  return (
-    <Pagination
-      onPageChange={pageNumber => console.log(pageNumber)}
-      aria-label="Pagination"
-      lastPage={lastPage}
-    >
-      <Pagination.Controls>
-        <Pagination.StepToPreviousButton aria-label="Previous" />
-        <Pagination.PageList>
-          {({state}) =>
-            state.range.map(pageNumber => (
-              <Pagination.PageListItem key={pageNumber}>
-                <Pagination.PageButton aria-label={`Page ${pageNumber}`} pageNumber={pageNumber} />
-              </Pagination.PageListItem>
-            ))
-          }
-        </Pagination.PageList>
-        <Pagination.StepToNextButton aria-label="Next" />
-      </Pagination.Controls>
-      <Pagination.AdditionalDetails shouldHideDetails>
-        {({state}) =>
-          `${getVisibleResultsMin(state.currentPage, resultCount)}-${getVisibleResultsMax(
-            state.currentPage,
-            resultCount,
-            totalCount
-          )} of ${totalCount} results`
-        }
-      </Pagination.AdditionalDetails>
-    </Pagination>
-  );
-};
-
-export const CustomRange = () => {
-  const resultCount = 10;
-  const totalCount = 100;
-  const lastPage = getLastPage(resultCount, totalCount);
-
-  return (
-    <Pagination
-      aria-label="Pagination"
-      lastPage={lastPage}
-      onPageChange={pageNumber => console.log(pageNumber)}
-      rangeSize={3}
-    >
-      <Pagination.Controls>
-        <Pagination.StepToPreviousButton aria-label="Previous" />
-        <Pagination.PageList>
-          {({state}) =>
-            state.range.map(pageNumber => (
-              <Pagination.PageListItem key={pageNumber}>
-                <Pagination.PageButton aria-label={`Page ${pageNumber}`} pageNumber={pageNumber} />
-              </Pagination.PageListItem>
-            ))
-          }
-        </Pagination.PageList>
-        <Pagination.StepToNextButton aria-label="Next" />
-      </Pagination.Controls>
-      <Pagination.AdditionalDetails shouldHideDetails>
-        {({state}) =>
-          `${getVisibleResultsMin(state.currentPage, resultCount)}-${getVisibleResultsMax(
-            state.currentPage,
-            resultCount,
-            totalCount
-          )} of ${totalCount} results`
-        }
-      </Pagination.AdditionalDetails>
-    </Pagination>
-  );
-};
-
-export const JumpControls = () => {
-  const resultCount = 10;
-  const totalCount = 100;
-  const lastPage = getLastPage(resultCount, totalCount);
-
-  return (
-    <Pagination
-      onPageChange={pageNumber => console.log(pageNumber)}
-      aria-label="Pagination"
-      lastPage={lastPage}
-    >
-      <Pagination.Controls>
-        <Pagination.JumpToFirstButton aria-label="First" />
-        <Pagination.StepToPreviousButton aria-label="Previous" />
-        <Pagination.PageList>
-          {({state}) =>
-            state.range.map(pageNumber => (
-              <Pagination.PageListItem key={pageNumber}>
-                <Pagination.PageButton aria-label={`Page ${pageNumber}`} pageNumber={pageNumber} />
-              </Pagination.PageListItem>
-            ))
-          }
-        </Pagination.PageList>
-        <Pagination.StepToNextButton aria-label="Next" />
-        <Pagination.JumpToLastButton aria-label="Last" />
-      </Pagination.Controls>
-      <Pagination.AdditionalDetails shouldHideDetails>
-        {({state}) =>
-          `${getVisibleResultsMin(state.currentPage, resultCount)}-${getVisibleResultsMax(
-            state.currentPage,
-            resultCount,
-            totalCount
-          )} of ${totalCount} results`
-        }
-      </Pagination.AdditionalDetails>
-    </Pagination>
-  );
-};
-
-export const GoToForm = () => {
-  const resultCount = 10;
-  const totalCount = 100;
-  const lastPage = getLastPage(resultCount, totalCount);
-
-  return (
-    <Pagination
-      onPageChange={pageNumber => console.log(pageNumber)}
-      aria-label="Pagination"
-      lastPage={lastPage}
-      rangeSize={3}
-    >
-      <Pagination.Controls>
-        <Pagination.JumpToFirstButton aria-label="First" />
-        <Pagination.StepToPreviousButton aria-label="Previous" />
-        <Pagination.PageList>
-          {({state}) =>
-            state.range.map(pageNumber => (
-              <Pagination.PageListItem key={pageNumber}>
-                <Pagination.PageButton aria-label={`Page ${pageNumber}`} pageNumber={pageNumber} />
-              </Pagination.PageListItem>
-            ))
-          }
-        </Pagination.PageList>
-        <Pagination.StepToNextButton aria-label="Next" />
-        <Pagination.JumpToLastButton aria-label="Last" />
-        <Pagination.GoToForm>
-          <Pagination.GoToTextInput />
-          <Pagination.GoToLabel>{({state}) => `of ${totalCount} results`}</Pagination.GoToLabel>
-        </Pagination.GoToForm>
-      </Pagination.Controls>
-      <Pagination.AdditionalDetails shouldHideDetails>
-        {({state}) =>
-          `${getVisibleResultsMin(state.currentPage, resultCount)}-${getVisibleResultsMax(
-            state.currentPage,
-            resultCount,
-            totalCount
-          )} of ${totalCount} results`
-        }
-      </Pagination.AdditionalDetails>
-    </Pagination>
-  );
-};
-
-export const ShowAdditionalDetails = () => {
-  const resultCount = 10;
-  const totalCount = 100;
-  const lastPage = getLastPage(resultCount, totalCount);
-
-  return (
-    <Pagination
-      onPageChange={pageNumber => console.log(pageNumber)}
-      aria-label="Pagination"
-      lastPage={lastPage}
-      rangeSize={3}
-      initialCurrentPage={3}
-    >
-      <Pagination.Controls>
-        <Pagination.JumpToFirstButton aria-label="First" />
-        <Pagination.StepToPreviousButton aria-label="Previous" />
-        <Pagination.PageList>
-          {({state}) =>
-            state.range.map(pageNumber => (
-              <Pagination.PageListItem key={pageNumber}>
-                <Pagination.PageButton aria-label={`Page ${pageNumber}`} pageNumber={pageNumber} />
-              </Pagination.PageListItem>
-            ))
-          }
-        </Pagination.PageList>
-        <Pagination.StepToNextButton aria-label="Next" />
-        <Pagination.JumpToLastButton aria-label="Last" />
-        <Pagination.GoToForm>
-          <Pagination.GoToTextInput />
-          <Pagination.GoToLabel>{() => `of ${totalCount} results`}</Pagination.GoToLabel>
-        </Pagination.GoToForm>
-      </Pagination.Controls>
-      <Pagination.AdditionalDetails>
-        {({state}) =>
-          `${getVisibleResultsMin(state.currentPage, resultCount)}-${getVisibleResultsMax(
-            state.currentPage,
-            resultCount,
-            totalCount
-          )} of ${totalCount} results`
-        }
-      </Pagination.AdditionalDetails>
-    </Pagination>
-  );
-};
-
-export const ExternalModel = () => {
-  const resultCount = 10;
-  const totalCount = 100;
-  const lastPage = getLastPage(resultCount, totalCount);
-
-  const model = usePaginationModel({
-    lastPage,
-    onPageChange: number => console.log(number),
-  });
-  return (
-    <Pagination aria-label="Pagination" model={model}>
-      <Pagination.Controls>
-        <Pagination.StepToPreviousButton aria-label="Previous" />
-        <Pagination.PageList>
-          {({state}) =>
-            state.range.map(pageNumber => (
-              <Pagination.PageListItem key={pageNumber}>
-                <Pagination.PageButton aria-label={`Page ${pageNumber}`} pageNumber={pageNumber} />
-              </Pagination.PageListItem>
-            ))
-          }
-        </Pagination.PageList>
-        <Pagination.StepToNextButton aria-label="Next" />
-      </Pagination.Controls>
-      <Pagination.AdditionalDetails shouldHideDetails>
-        {({state}) =>
-          `${getVisibleResultsMin(state.currentPage, resultCount)}-${getVisibleResultsMax(
-            state.currentPage,
-            resultCount,
-            totalCount
-          )} of ${totalCount} results`
-        }
-      </Pagination.AdditionalDetails>
-    </Pagination>
-  );
-};
-
-export const RTL = () => {
-  const resultCount = 10;
-  const totalCount = 100;
-  const lastPage = getLastPage(resultCount, totalCount);
-
-  return (
-    <CanvasProvider theme={{canvas: {direction: ContentDirection.RTL}}}>
-      <Pagination aria-label="Pagination" lastPage={lastPage}>
-        <Pagination.Controls>
-          <Pagination.JumpToFirstButton aria-label="First" />
-          <Pagination.StepToPreviousButton aria-label="Previous" />
-          <Pagination.PageList>
-            {({state}) =>
-              state.range.map(pageNumber => (
-                <Pagination.PageListItem key={pageNumber}>
-                  <Pagination.PageButton
-                    aria-label={`Page ${pageNumber}`}
-                    pageNumber={pageNumber}
-                  />
-                </Pagination.PageListItem>
-              ))
-            }
-          </Pagination.PageList>
-          <Pagination.StepToNextButton aria-label="Next" />
-          <Pagination.JumpToLastButton aria-label="Last" />
-          <Pagination.GoToForm>
-            <Pagination.GoToTextInput />
-            <Pagination.GoToLabel>{({state}) => `من 100 صفحات`}</Pagination.GoToLabel>
-          </Pagination.GoToForm>
-        </Pagination.Controls>
-        <Pagination.AdditionalDetails shouldHideDetails>
-          {({state}) =>
-            `${getVisibleResultsMax(
-              state.currentPage,
-              resultCount,
-              totalCount
-            )}-${getVisibleResultsMin(state.currentPage, resultCount)} من 100 صفحات`
-          }
-        </Pagination.AdditionalDetails>
-      </Pagination>
-    </CanvasProvider>
-  );
-};
